fix: #1676 cache updates
* fixes as per code review

* ordering by date, not IDs

* better default parameters

* better fixture location
