+++ packages/server/__tests__/arpa_reporter/server/fixtures/fixtures.js
@@ -103,12 +103,196 @@ const uploads = {
     },
 };
 
+const audit_report_data = {
+    obligations: [{
+        'Reporting Period': 'Quarterly 1',
+        'Period End Date': new Date(2021, 11, 31),
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/601a2011-91d5-4acb-b83e-f47ee8ae462f\',\'UPLOAD_CLEAN_ID6.xlsm\')',
+        },
+        'Adopted Budget (EC tabs)': 300000,
+        'Total Cumulative Obligations (EC tabs)': 300000,
+        'Total Cumulative Expenditures (EC tabs)': 150000,
+        'Current Period Obligations (EC tabs)': 150000,
+        'Current Period Expenditures (EC tabs)': 150000,
+        'Subaward Obligations (Subaward >50k)': 300000,
+        'Total Expenditure Amount (Expenditures >50k)': 150000,
+        'Current Period Obligations (Aggregate Awards <50k)': 0,
+        'Current Period Expenditures (Aggregate Awards <50k)': 0,
+    }, {
+        'Reporting Period': 'Quarterly 2',
+        'Period End Date': new Date(2022, 2, 31),
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/3bb6dbb3-741c-43d1-957a-2d35dd1dd2a9\',\'UPLOAD_CLEAN_ID7.xlsm\')',
+        },
+        'Adopted Budget (EC tabs)': 300000,
+        'Total Cumulative Obligations (EC tabs)': 300000,
+        'Total Cumulative Expenditures (EC tabs)': 150000,
+        'Current Period Obligations (EC tabs)': 150000,
+        'Current Period Expenditures (EC tabs)': 150000,
+        'Subaward Obligations (Subaward >50k)': 300000,
+        'Total Expenditure Amount (Expenditures >50k)': 150000,
+        'Current Period Obligations (Aggregate Awards <50k)': 0,
+        'Current Period Expenditures (Aggregate Awards <50k)': 0,
+    }, {
+        'Reporting Period': 'Quarterly 3',
+        'Period End Date': new Date(2022, 5, 30),
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/db5de7d3-4209-46d0-9479-8696491aadfc\',\'UPLOAD_CLEAN_ID8.xlsm\')',
+        },
+        'Adopted Budget (EC tabs)': 300000,
+        'Total Cumulative Obligations (EC tabs)': 300000,
+        'Total Cumulative Expenditures (EC tabs)': 150000,
+        'Current Period Obligations (EC tabs)': 150000,
+        'Current Period Expenditures (EC tabs)': 150000,
+        'Subaward Obligations (Subaward >50k)': 300000,
+        'Total Expenditure Amount (Expenditures >50k)': 150000,
+        'Current Period Obligations (Aggregate Awards <50k)': 0,
+        'Current Period Expenditures (Aggregate Awards <50k)': 0,
+    }],
+    projectSummaries: [{
+        'Project ID': 6,
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/601a2011-91d5-4acb-b83e-f47ee8ae462f\',\'UPLOAD_CLEAN_ID6.xlsm\')',
+        },
+        'Last Reported': 'Quarterly 1',
+        'Adopted Budget': 300000,
+        'Total Cumulative Obligations': 300000,
+        'Total Cumulative Expenditures': 150000,
+        'Current Period Obligations': 150000,
+        'Current Period Expenditures': 150000,
+        'Completion Status': 'Completed less than 50%',
+    }, {
+        'Project ID': 7,
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/3bb6dbb3-741c-43d1-957a-2d35dd1dd2a9\',\'UPLOAD_CLEAN_ID7.xlsm\')',
+        },
+        'Last Reported': 'Quarterly 2',
+        'Adopted Budget': 300000,
+        'Total Cumulative Obligations': 300000,
+        'Total Cumulative Expenditures': 150000,
+        'Current Period Obligations': 150000,
+        'Current Period Expenditures': 150000,
+        'Completion Status': 'Completed less than 50%',
+    },
+    {
+        'Project ID': 8,
+        Upload: {
+            f: '=HYPERLINK(\'http://localhost:8080/arpa_reporter/uploads/db5de7d3-4209-46d0-9479-8696491aadfc\',\'UPLOAD_CLEAN_ID8.xlsm\')',
+        },
+        'Last Reported': 'Quarterly 3',
+        'Adopted Budget': 300000,
+        'Total Cumulative Obligations': 300000,
+        'Total Cumulative Expenditures': 150000,
+        'Current Period Obligations': 150000,
+        'Current Period Expenditures': 150000,
+        'Completion Status': 'Completed less than 50%',
+    }],
+    projectSummaryGroupedByProject: [{
+        'Project ID': '6',
+        'Project Description': 'This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state. \'New Family Child Care Providers\' can be defined, for the purposes of this program, as providers who do not currently hold a license with the Department of Human Services, which may include providers who previously held licenses in good standing with the Department or providers new to the field entirely. This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state.',
+        'Project Expenditure Category Group': '2-Negative Economic Impacts',
+        'Project Expenditure Category': '2.32-Business Incubators and Start-Up or Expansion Assistance',
+        '2021-12-31 Total Aggregate Expenditures': 150000,
+        '2021-12-31 Total Expenditures for Awards Greater or Equal to $50k': 0,
+        '2021-12-31 Total Aggregate Obligations': 300000,
+        '2021-12-31 Total Obligations for Awards Greater or Equal to $50k': 0,
+        'Capital Expenditure Amount': 0,
+    }, {
+        'Project ID': '7',
+        'Project Description': 'This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state. \'New Family Child Care Providers\' can be defined, for the purposes of this program, as providers who do not currently hold a license with the Department of Human Services, which may include providers who previously held licenses in good standing with the Department or providers new to the field entirely. This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state.',
+        'Project Expenditure Category Group': '2-Negative Economic Impacts',
+        'Project Expenditure Category': '2.32-Business Incubators and Start-Up or Expansion Assistance',
+        '2022-03-31 Total Aggregate Expenditures': 150000,
+        '2022-03-31 Total Expenditures for Awards Greater or Equal to $50k': 0,
+        '2022-03-31 Total Aggregate Obligations': 300000,
+        '2022-03-31 Total Obligations for Awards Greater or Equal to $50k': 0,
+        'Capital Expenditure Amount': 0,
+    }, {
+        'Project ID': '8',
+        'Project Description': 'This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state. \'New Family Child Care Providers\' can be defined, for the purposes of this program, as providers who do not currently hold a license with the Department of Human Services, which may include providers who previously held licenses in good standing with the Department or providers new to the field entirely. This project will fund the start-up costs of new Family Child Care providers to open high-quality FCC options and increase the overall supply of child care in the state.',
+        'Project Expenditure Category Group': '2-Negative Economic Impacts',
+        'Project Expenditure Category': '2.32-Business Incubators and Start-Up or Expansion Assistance',
+        '2022-06-30 Total Aggregate Expenditures': 150000,
+        '2022-06-30 Total Expenditures for Awards Greater or Equal to $50k': 0,
+        '2022-06-30 Total Aggregate Obligations': 300000,
+        '2022-06-30 Total Obligations for Awards Greater or Equal to $50k': 0,
+        'Capital Expenditure Amount': 0,
+    }],
+    KPIDataGroupedByProject: [{
+        'Project ID': '6',
+        'Number of Subawards': 0,
+        'Number of Expenditures': 1,
+        'Evidence Based Total Spend': 0,
+        'Evidence based total spend': null,
+    }, {
+        'Project ID': '7',
+        'Number of Subawards': 0,
+        'Number of Expenditures': 1,
+        'Evidence Based Total Spend': 0,
+        'Evidence based total spend': null,
+    }, {
+        'Project ID': '8',
+        'Number of Subawards': 0,
+        'Number of Expenditures': 1,
+        'Evidence Based Total Spend': 0,
+        'Evidence based total spend': null,
+    }],
+};
+
+const session = {
+    user: {
+        id: 1,
+        email: 'alex@usdigitalresponse.org',
+        name: 'Alex Allain',
+        role_id: 1,
+        role_name: 'admin',
+        role_rules: {},
+        agency_id: 0,
+        agency_name: 'USDR',
+        agency_abbreviation: 'USDR',
+        agency_parent_id_id: null,
+        agency_warning_threshold: 30,
+        agency_danger_threshold: 15,
+        tenant_id: 1,
+        tenant_display_name: 'USDR Tenant',
+        tenant_main_agency_id: 400,
+        tenant_uses_spoc_process: false,
+        tags: null,
+        role: { id: 1, name: 'admin', rules: {} },
+        agency: {
+            id: 0,
+            name: 'USDR',
+            abbreviation: 'USDR',
+            agency_parent_id: undefined,
+            warning_threshold: 30,
+            danger_threshold: 15,
+            main_agency_id: undefined,
+            subagencies: [],
+        },
+        tenant: {
+            id: 1,
+            display_name: 'USDR Tenant',
+            main_agency_id: 400,
+            uses_spoc_process: false,
+        },
+        emailPreferences: {
+            GRANT_ASSIGNMENT: 'SUBSCRIBED',
+            GRANT_INTEREST: 'SUBSCRIBED',
+            GRANT_DIGEST: 'SUBSCRIBED',
+        },
+    },
+    selectedAgency: 0,
+};
+
 module.exports = {
     TABLES,
     reportingPeriods,
     uploads,
     TENANT_ID,
     users,
+    audit_report_data,
+    session,
 };
 
 module.exports.clean = async (knex) => {
